refix previous fix to Monitor
The previous fix to Monitor only accounted for threads that were parked in
between attempts to take the lock; they didn't also account for the loss of
a thread while trying to take the lock.  If the thread that holds the lock
is also waiting on some condition to be set by an incoming activity, and
there are no threads to handle the message, this could lead to a deadlock.
Expand the scope of increasing parallelism to include all statements
that may block for an extended period.