change interface to acceptGradients to include multiplier
NativeLearner::acceptGradients previously accepted a parameter numMB,
which was used to discount the gradients by a factor of 1.0/numMB.
Changed the interface of acceptGradients so that the multipler itself
(1.0/numMB) is passed.  Also changed the learner implementations to
roll the multiplier into the update, to save iterating over all gradients
once before actually applying the update.